+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Split `Iterable.zip()` into `Interable.zip()`, `Iterable.zipPartial()` and `Iterable.zipPartialWith()`.
 * Fix a bug in `Ordering.nullsFirst` and `Ordering.nullsLast`.
 * Fix a race condition in cache value resolution.
+* Fix a rounding bug of numeric formats with precision 0.
+* Add human number printer (1 kilo, 1 mega, ...).
 * Add a map function to `Tuple` classes.
 
 ## 2.7.0